+++ lib/tinysky/lexicon.rb
@@ -1,5 +1,6 @@
 module Tinysky
   module Lexicon
+    EXTERNAL_EMBED = "app.bsky.embed.external"
     FEED_POST = "app.bsky.feed.post"
     LINK_FACET = "app.bsky.richtext.facet#link"
   end