+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog][] and this project adheres to
 
 * Add update script ([#7][])
 * Add support for link facets ([#15][])
+* Add support for record embeds ([#16][])
 
 ### Changed
 
@@ -71,3 +72,4 @@ The format is based on [Keep a Changelog][] and this project adheres to
 [#13]: https://github.com/jonallured/tinysky/pull/13
 [#14]: https://github.com/jonallured/tinysky/pull/14
 [#15]: https://github.com/jonallured/tinysky/pull/15
+[#16]: https://github.com/jonallured/tinysky/pull/16